rearrange and simplify things from /cli to /create-woo-extension
