--- packages/js/cli/.gitignore
@@ -1,3 +1,0 @@
-vendor
-node_modules
-.turbo
