--- packages/js/cli/assets/new/template/.editorconfig.mustache
@@ -1,27 +1,0 @@
-# This file is for unifying the coding style for different editors and IDEs
-# editorconfig.org
-
-# WordPress Coding Standards
-# https://make.wordpress.org/core/handbook/coding-standards/
-
-root = true
-
-[*]
-charset = utf-8
-end_of_line = lf
-indent_size = 4
-tab_width = 4
-indent_style = tab
-insert_final_newline = true
-trim_trailing_whitespace = true
-
-[*.txt]
-trim_trailing_whitespace = false
-
-[*.{md,json,yml}]
-trim_trailing_whitespace = false
-indent_style = space
-indent_size = 2
-
-[*.json]
-indent_style = tab
