--- packages/js/cli/assets/new/template/.eslintrc.js.mustache
@@ -1,6 +1,0 @@
-module.exports = {
-	extends: [ 'plugin:@woocommerce/eslint-plugin/recommended' ],
-	rules: {
-		'react/react-in-jsx-scope': 'off',
-	},
-};
