--- packages/js/cli/assets/new/template/.gitignore.mustache
@@ -1,62 +1,0 @@
-
-# Operating System files
-.DS_Store
-Thumbs.db
-
-# IDE files
-.idea
-.vscode/*
-project.xml
-project.properties
-.project
-.settings*
-*.sublime-project
-*.sublime-workspace
-.sublimelinterrc
-
-# Sass
-.sass-cache/
-
-# Logs
-logs/
-
-# Environment files
-wp-cli.local.yml
-yarn-error.log
-npm-debug.log
-.pnpm-debug.log
-
-# Build files
-*.sql
-*.swp
-*.zip
-
-# Built packages
-build/
-build-module/
-build-style/
-build-types/
-dist/
-
-# Project files
-node_modules/
-vendor/
-
-# TypeScript files
-tsconfig.tsbuildinfo
-
-# Node Package Dependencies
-package-lock.json
-
-# wp-env config
-.wp-env.override.json
-
-# Unit tests
-tmp/
-
-# Composer
-vendor/
-bin/composer/**/vendor/
-lib/vendor/
-contributors.md
-contributors.html
