--- packages/js/cli/assets/new/template/.prettierrc.json.mustache
@@ -1,1 +1,0 @@
-"@wordpress/prettier-config"
