--- packages/js/cli/assets/new/template/.wp-env.json.mustache
@@ -1,13 +1,0 @@
-{
-	"phpVersion": null,
-	"core": null,
-	"plugins": [
-		"https://downloads.wordpress.org/plugin/woocommerce.7.3.0.zip",
-		"."
-	],
-	"config": {
-		"JETPACK_AUTOLOAD_DEV": true,
-		"WP_DEBUG": true,
-		"SCRIPT_DEBUG": true
-	}
-}
