--- packages/js/cli/assets/new/template/composer.json.mustache
@@ -1,18 +1,0 @@
-{
-	"name": "{{namespace}}/{{slug}}",
-	"description": "WooCommerce plugin",
-	"autoload": {
-		"psr-4": {
-			"{{slugPascalCase}}\\": "includes/"
-		}
-	},
-	"require-dev": {
-		"phpunit/phpunit": "^9.0",
-		"automattic/jetpack-autoloader": "^2"
-	},
-	"config": {
-		"allow-plugins": {
-			"automattic/jetpack-autoloader": true
-		}
-	}
-}
