--- packages/js/cli/assets/new/template/languages/woo-plugin-setup.pot.mustache
@@ -1,45 +1,0 @@
-# Copyright (C) 2022 WooCommerce
-# This file is distributed under the GNU General Public License v3.0.
-msgid ""
-msgstr ""
-"Project-Id-Version: Woo Plugin Setup 1.0.0\n"
-"Report-Msgid-Bugs-To: https://wordpress.org/support/plugin/{{slug}}\n"
-"Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
-"Language-Team: LANGUAGE <LL@li.org>\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"POT-Creation-Date: 2022-11-27T21:35:18+00:00\n"
-"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
-"X-Generator: WP-CLI 2.7.1\n"
-"X-Domain: {{slug}}\n"
-
-#. Plugin Name of the plugin
-msgid "Woo Plugin Setup"
-msgstr ""
-
-#. Plugin URI of the plugin
-msgid "https://github.com/psealock/woo-plugin-setup"
-msgstr ""
-
-#. Description of the plugin
-msgid "This is an example of an extension template to follow."
-msgstr ""
-
-#. Author of the plugin
-msgid "WooCommerce"
-msgstr ""
-
-#. Author URI of the plugin
-msgid "https://woocommerce.com"
-msgstr ""
-
-#: build/index.js:1
-#: src/index.js:13
-msgid "My Example Extension"
-msgstr ""
-
-#: build/index.js:1
-#: src/index.js:20
-msgid "My Example Page"
-msgstr ""
