--- packages/js/cli/assets/new/template/src/index.scss.mustache
@@ -1,3 +1,0 @@
-h1 {
-	background-color: gold;
-}
