--- packages/js/cli/assets/new/template/tests/Test.php.mustache
@@ -1,15 +1,0 @@
-<?php
-use PHPUnit\Framework\TestCase;
-
-/**
- * Basic unit tests.
- */
-final class Test extends TestCase {
-
-	/**
-	 * Test if 1 === 1.
-	 */
-	public function testEquality(): void {
-		$this->assertEquals( 1, 1 );
-	}
-}
