--- packages/js/cli/changelog/add-woo-cli
@@ -1,4 +1,0 @@
-Significance: major
-Type: add
-
-Add commander.js-based wrapper for scaffolding new WooCommerce extensions.
