--- packages/js/cli/composer.json
@@ -1,32 +1,0 @@
-{
-	"name": "woocommerce/cli",
-	"description": "A command line interface for developing WooCommerce extensions.",
-	"type": "project",
-	"license": "GPL-3.0-or-later",
-	"minimum-stability": "dev",
-	"require-dev": {
-		"automattic/jetpack-changelogger": "3.3.0"
-	},
-	"config": {
-		"platform": {
-			"php": "7.2"
-		}
-	},
-	"extra": {
-		"changelogger": {
-			"formatter": {
-				"filename": "../../../tools/changelogger/class-package-formatter.php"
-			},
-			"types": {
-				"fix": "Fixes an existing bug",
-				"add": "Adds functionality",
-				"update": "Update existing functionality",
-				"dev": "Development related task",
-				"tweak": "A minor adjustment to the codebase",
-				"performance": "Address performance issues",
-				"enhancement": "Improve existing functionality"
-			},
-			"changelog": "CHANGELOG.md"
-		}
-	}
-}
