--- packages/js/cli/package.json
@@ -1,41 +1,0 @@
-{
-  "name": "@woocommerce/cli",
-  "version": "0.1.0",
-  "description": "A command line interface for developing WooCommerce extensions.",
-  "author": "Automattic",
-  "license": "GPL-3.0-or-later",
-  "keywords": ["woocommerce", "cli", "wordpress"],
-  "main": "index.js",
-  "bin": {
-    "woo": "./index.js"
-  },
-  "scripts": {
-    "changelog": "composer exec -- changelogger",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "require-dev": {
-		"automattic/jetpack-changelogger": "3.3.0"
-	},
-  "dependencies": {
-    "@wordpress/create-block": "^4.11.0",
-    "commander": "^10.0.0",
-    "slug": "^8.2.2"
-  },
-	"extra": {
-		"changelogger": {
-			"formatter": {
-				"filename": "../../../tools/changelogger/class-package-formatter.php"
-			},
-			"types": {
-				"fix": "Fixes an existing bug",
-				"add": "Adds functionality",
-				"update": "Update existing functionality",
-				"dev": "Development related task",
-				"tweak": "A minor adjustment to the codebase",
-				"performance": "Address performance issues",
-				"enhancement": "Improve existing functionality"
-			},
-			"changelog": "CHANGELOG.md"
-		}
-	}
-}
