+++ packages/js/create-woo-extension/lib/package-info.js
