+++ packages/js/create-woo-extension/template/$slug.php.mustache
