+++ packages/js/create-woo-extension/template/.editorconfig.mustache
