+++ packages/js/create-woo-extension/template/.eslintrc.js.mustache
