+++ packages/js/create-woo-extension/template/.gitignore.mustache
