+++ packages/js/create-woo-extension/template/.prettierrc.json.mustache
