+++ packages/js/create-woo-extension/template/.wp-env.json.mustache
