+++ packages/js/create-woo-extension/template/README.md.mustache
