+++ packages/js/create-woo-extension/template/composer.json.mustache
