+++ packages/js/create-woo-extension/template/includes/admin/setup.php.mustache
