+++ packages/js/create-woo-extension/template/index.js
