+++ packages/js/create-woo-extension/template/languages/woo-plugin-setup.pot.mustache
