+++ packages/js/create-woo-extension/template/package.json
