+++ packages/js/create-woo-extension/template/src/index.js.mustache
