+++ packages/js/create-woo-extension/template/src/index.scss.mustache
