+++ packages/js/create-woo-extension/template/tests/Test.php.mustache
