+++ packages/js/create-woo-extension/template/webpack.config.js.mustache
